requirements.txt
mcrcon==0.7.0
PyYAML==6.0.1
